% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateIntervals.R
\name{estimateTimePoint}
\alias{estimateTimePoint}
\title{Function to estimate isotopic value for specific time point(s)}
\usage{
estimateTimePoint(object, time, intervalProb = 0.95)
}
\arguments{
\item{object}{An object of class \link{TemporalIso} fitted with estimateIntervals()}

\item{time}{A vector of numerics, indicating the point in time to extract estimates}

\item{intervalProb}{A numeric between 0 and 1 indicating the coverage of the credible interval}
}
\value{
A data.frame with time points and their means, standard deviations and credible intervals
}
\description{
Once the isotopic values in time course are estimated along
with estimateIntervals(), this function provides an interface to extract the mean, standard deviation and quantiles for a specific point in time
}
