% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\docType{methods}
\name{export}
\alias{export}
\alias{export,TemporalIso-method}
\title{Export of TemporalIso samples as csv}
\usage{
\S4method{export}{TemporalIso}(fit, file, interval = NULL, ...)
}
\arguments{
\item{fit}{TemporalIso: An instance of class TemporalIso,
generated from \code{\link{estimateIntervals}}}

\item{file}{character: a character string specifying the path to a .csv-file.}

\item{interval}{numeric values which are part of the specified time intervals.
Defaults to all intervals being exported}

\item{...}{Optional arguments to \link[utils]{write.csv}.
#'}
}
\description{
Writes the interval samples of a TemporalIso-object to a .csv-File.
}
\seealso{
\link[utils]{write.csv}
}
