% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{plot,TemporalIso,missing-method}
\alias{plot,TemporalIso,missing-method}
\title{Plots for TemporalIso objects}
\usage{
\S4method{plot}{TemporalIso,missing}(x, ..., plotfun = "isotopic")
}
\arguments{
\item{x}{TemporalIso: An instance of class TemporalIso,
generated from \code{\link{estimateIntervals}}}

\item{...}{Optional arguments to the plotting functions.}

\item{plotfun}{character: a character string specifying which plotting function is to be used. The
default is "isotopic", whis is a \code{\link[rstan]{stan_plot}} only for the interval values.
"stan_plot", "stan_trace", "stan_dens" etc regress to the functionality of rstan.}
}
\value{
a ggplot object which can be further customized using the ggplot2 package.
}
\description{
Inherited from \link[rstan]{rstan-plotting-functions}, the default plot shows
posterior uncertainty intervals and point estimates of isotopic values for each interval. The
plot method can also regress to standard rstan methods.
}
\seealso{
\link[rstan]{rstan-plotting-functions}, \link[rstan]{rstan_gg_options}
}
