% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTime}
\alias{plotTime}
\title{Plot of credibility intervals for each time interval, plotted as timeseries}
\usage{
plotTime(object, prop = 0.8, plotShifts = FALSE, ...)
}
\arguments{
\item{object}{model of class \code{\link{TemporalIso}}}

\item{prop}{double between 0 and 1: length of credibility interval. The default value is 80 percent.}

\item{plotShifts}{boolean if shifts shall be marked or not. Defaults to False.}

\item{...}{arguments handed to \code{\link{getShiftTime}}}
}
\value{
a \link[ggplot2]{ggplot} object.
}
\description{
The function plots the credibility intervals for each
time interval and returns a ggplot object, which is further customizable.
}
