% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{TemporalIso-class}
\alias{TemporalIso-class}
\alias{TemporalIso}
\title{S4 class for temporal estimation of isotopic values}
\description{
Inherits from \code{\link[rstan]{stanfit}}
}
\section{Slots}{

\describe{
\item{\code{regressor}}{regressor used for estimation (class matrix)}

\item{\code{time}}{timestamps (class vector)}
}}

