% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMESPACE.R
\docType{package}
\name{OsteoBioR-package}
\alias{OsteoBioR-package}
\alias{OsteoBioR}
\title{The 'OsteoBioR' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version NA. http://mc-stan.org
}
